Proper character escaping.
Character escaping consists in converting "problematic" characters
(characters LaTeX treats differently) into a sequence of characters that
would no bother LaTeX and yet be translated into something whose
representation would close or equal to what the user intended in
his/hers yml code.

The old character escaping code used replace and a dict with a
conversion table.  Well, this would work as expected as long as no
character that needed conversion appeared as part of a translated code
for other character, in which case it risked suffering double-conversion
and outputting the wrong escape code.

Take '\' and '}'. The former would be translated into "\textbackslash"
and the later into "\{". But since dict iterators do not guarantee
iteration ordering, even though "\" was added to the dict before "}", it
may well be the case that in the first pass we translate all "}" into
"\}" and than, unexpectedly, converted them into "\textbackslash }" --
what, unfortunately, is not the right escaping for "}" and will result
into a LaTeX syntax error.

So, in order to do this correctly, one has to convert every character in
the input sequence individually. This will take O(n)*O(m) worse case,
where m is the size of the conversion table -- the same complexity of
the original code. Since python dictionaries are pretty optimized, the
average case can be seen as O(n).

( Damn... such a long commit message for such a small patch)